Move MLSwitch to new directory
